MetaDataContainer send on viewer add implementation + some related changes and convenience changes

+++ src/de/atlasmc/util/EulerAngle.java
@@ -18,15 +18,15 @@ public class EulerAngle {
 		this.z = angle.z;
 	}
 
-	public double getX() {
+	public float getX() {
 		return x;
 	}
 	
-	public double getY() {
+	public float getY() {
 		return y;
 	}
 	
-	public double getZ() {
+	public float getZ() {
 		return z;
 	}
